project/backend/requirements.txt
psycopg2-binary
psycopg
django
djangorestframework
django-cors-headers
